+++ src/Popover/Popover.js
@@ -2,6 +2,7 @@ import chain from 'chain-function';
 import classnames from 'classnames';
 import { findDOMNode } from 'react-dom';
 import FocusManager from '../utils/focusManager/focusManager';
+import keycode from 'keycode';
 import Popper from '../utils/_Popper';
 import { POPPER_PLACEMENTS } from '../utils/constants';
 import PropTypes from 'prop-types';
@@ -18,6 +19,18 @@ class Popover extends Component {
         };
     }
 
+    isButton = (node) => {
+        if (!node) {
+            return false;
+        }
+        if (typeof node.type === 'string') {
+            return node.type.toLowerCase() === 'button';
+        } else if (node.type.displayName) {
+            return node.type.displayName.toLowerCase() === 'button';
+        }
+        return false;
+    };
+
     triggerBody = () => {
         if (!this.props.disabled) {
             this.setState(prevState => ({
@@ -28,7 +41,7 @@ class Popover extends Component {
 
     handleFocusManager = () => {
         if (this.state.isExpanded && this.popover) {
-            this.focusManager = new FocusManager(this.popover, this.props.useArrowKeyNavigation);
+            this.focusManager = new FocusManager(this.popover, this.controlRef, this.props.useArrowKeyNavigation);
         }
     }
 
@@ -37,21 +50,39 @@ class Popover extends Component {
             this.setState({
                 isExpanded: false
             });
-
-            if (this.controlRef) {
-                if (tabbable.isTabbable(this.controlRef)) {
-                    this.controlRef.focus();
-                } else {
-                    const firstTabbableNode = tabbable(this.controlRef)[0];
-                    firstTabbableNode && firstTabbableNode.focus();
-                }
-            }
         }
     };
+
+    handleEscapeKey = () => {
+        this.handleOutsideClick();
+
+        if (this.controlRef) {
+            if (tabbable.isTabbable(this.controlRef)) {
+                this.controlRef.focus();
+            } else {
+                const firstTabbableNode = tabbable(this.controlRef)[0];
+                firstTabbableNode && firstTabbableNode.focus();
+            }
+        }
+    }
+
+    handleKeyPress = (event, node, onClickFunctions) => {
+        if (!this.isButton(node)) {
+            switch (keycode(event)) {
+                case 'enter':
+                case 'space':
+                    event.preventDefault();
+                    onClickFunctions();
+                    break;
+                default:
+            }
+        }
+    }
 
     render() {
         const {
             disableEdgeDetection,
+            disableKeyPressHandler,
             disableStyles,
             onClickOutside,
             onEscapeKey,
@@ -71,17 +102,29 @@ class Popover extends Component {
             onClickFunctions = chain(this.triggerBody, control.props.onClick);
         }
 
+        let controlProps = {
+            onClick: onClickFunctions,
+            ref: (c) => {
+                this.controlRef = findDOMNode(c);
+            }
+        };
+
         const innerRef = (c) => {
             this.popover = findDOMNode(c);
             this.handleFocusManager();
         };
 
-        const referenceComponent = React.cloneElement(control, {
-            onClick: onClickFunctions,
-            ref: (c) => {
-                this.controlRef = findDOMNode(c);
-            }
-        });
+        if (!disableKeyPressHandler) {
+            controlProps = {
+                ...controlProps,
+                tabIndex: 0,
+                role: 'button',
+                'aria-haspopup': true,
+                onKeyPress: (event) => this.handleKeyPress(event, control, onClickFunctions)
+            };
+        }
+
+        const referenceComponent = React.cloneElement(control, controlProps);
 
         const popoverClasses = classnames('fd-popover', className);
 
@@ -93,7 +136,7 @@ class Popover extends Component {
                     innerRef={innerRef}
                     noArrow={noArrow}
                     onClickOutside={chain(this.handleOutsideClick, onClickOutside)}
-                    onEscapeKey={chain(this.handleOutsideClick, onEscapeKey)}
+                    onEscapeKey={chain(this.handleEscapeKey, onEscapeKey)}
                     popperPlacement={placement}
                     popperProps={popperProps}
                     referenceClassName='fd-popover__control'
@@ -116,6 +159,7 @@ Popover.propTypes = {
     customStyles: PropTypes.object,
     disabled: PropTypes.bool,
     disableEdgeDetection: PropTypes.bool,
+    disableKeyPressHandler: PropTypes.bool,
     disableStyles: PropTypes.bool,
     noArrow: PropTypes.bool,
     placement: PropTypes.oneOf(POPPER_PLACEMENTS),
@@ -134,6 +178,7 @@ Popover.propDescriptions = {
     body: 'Node(s) to render in the overlay.',
     control: 'Node to render as the reference element (that the `body` will be placed in relation to).',
     disableEdgeDetection: 'Set to **true** to render popover without edge detection so popover will not flip from top to bottom with scroll.',
+    disableKeyPressHandler: 'Set to **true** to remove onKeyPress handler and aria-* roles. Only do so if the control is a complex component such as a FormInput with Button.',
     noArrow: 'Set to **true** to render a popover without an arrow.',
     placement: 'Initial position of the `body` (overlay) related to the `control`.',
     popperProps: 'Additional props to be spread to the overlay element, supported by <a href="https://popper.js.org" target="_blank">popper.js</a>.',