+++ src/utils/focusManager/focusManager.js
@@ -1,9 +1,12 @@
 import keycode from 'keycode';
 import tabbable from 'tabbable';
+/* eslint-disable */
+
 
 export default class FocusManager {
-    constructor(trapNode, useArrowKeys = false) {
+    constructor(trapNode, controlNode, useArrowKeys = false) {
         this.container = trapNode;
+        this.firstOuterTabbableNode = tabbable.isTabbable(controlNode) ? controlNode : tabbable(controlNode)[0];
         this.tabbableNodes = tabbable(this.container);
         this.useArrowKeys = useArrowKeys;
 
@@ -12,20 +15,6 @@ export default class FocusManager {
 
     isFocusContained = (e) => {
         return (e.target === window && this.container && !this.container.contains(document.activeElement));
-    }
-
-    findParentTabbableElement = (target) => {
-        const index = this.tabbableNodes.indexOf(target);
-
-        if (index >= 0) {
-            return index;
-        } else {
-            if (target.parentNode) {
-                return this.findParentTabbableElement(target.parentNode);
-            } else {
-                return -1;
-            }
-        }
     }
 
     keyHandler = (e) => {
@@ -48,7 +37,7 @@ export default class FocusManager {
             }
 
             this.tabbableNodes = tabbable(this.container);
-            const currentIndex = this.findParentTabbableElement(e.target);
+            const currentIndex = this.tabbableNodes.indexOf(e.target);
             const lastNode = this.tabbableNodes[this.tabbableNodes.length - 1];
             const firstNode = this.tabbableNodes[0];
 
@@ -65,6 +54,13 @@ export default class FocusManager {
                     this.tryFocus(this.tabbableNodes[currentIndex + 1]);
                 }
             }
+        } else if (this.useArrowKeys && e.keyCode === keycode.codes.tab) { 
+            // navigate out of component with tab when arrow-key navigation enabled
+            e.preventDefault();
+
+            const documentTabbableElements = tabbable(document);
+            const nextElementIndex = documentTabbableElements.indexOf(this.firstOuterTabbableNode) + (e.shiftKey ? -1 : 1);
+            this.tryFocus(documentTabbableElements[nextElementIndex]);
         }
     };