+++ src/utils/focusManager/focusManager.test.js
@@ -10,6 +10,9 @@ Object.defineProperty(HTMLElement.prototype, 'offsetParent', {
 });
 
 describe('focus manager', () => {
+    let trapNode;
+    let controlNode;
+
     beforeAll(() => {
         global.document.createElement = function() {
             const [type] = arguments;
@@ -25,38 +28,36 @@ describe('focus manager', () => {
         global.document.createElement = ce;
     });
 
-    const createNode = () => {
-        const div = document.createElement('div');
+    beforeEach(() => {
+        trapNode = document.createElement('div');
+        controlNode = document.createElement('button');
         const a1 = document.createElement('a');
         const a2 = document.createElement('a');
         const a3 = document.createElement('a');
         const a4 = document.createElement('a');
         const span = document.createElement('span');
 
-        div.id = 'mainDiv';
-
-        div.appendChild(a1);
-        div.appendChild(a2);
-        div.appendChild(span);
-        div.appendChild(a3);
+        trapNode.appendChild(a1);
+        trapNode.appendChild(a2);
+        trapNode.appendChild(span);
+        trapNode.appendChild(a3);
 
         span.appendChild(a4);
-        return div;
-    };
+        controlNode.appendChild(trapNode);
+    });
 
     describe('Default Behavior', () => {
         it('should have tabbable elements', () => {
-            const manager = new FocusManager(createNode());
+            const manager = new FocusManager(trapNode, controlNode);
 
             expect(manager.tabbableNodes.length).toEqual(4);
         });
 
         it('should focus on nodes via tryFocus', () => {
-            const node = createNode();
-            const manager = new FocusManager(node);
-            const link = node.querySelector('a');
+            const manager = new FocusManager(trapNode, controlNode);
+            const link = trapNode.querySelector('a');
 
-            manager.tryFocus(node);
+            manager.tryFocus(link);
 
             expect(document.activeElement).toEqual(link);
         });
@@ -64,10 +65,9 @@ describe('focus manager', () => {
 
     describe('Event Behavior', () => {
         it('should handle tabbing back 1 element', () => {
-            const node = createNode();
-            const manager = new FocusManager(node);
+            const manager = new FocusManager(trapNode, controlNode);
 
-            const anchors = node.querySelectorAll('a');
+            const anchors = trapNode.querySelectorAll('a');
             const a1 = anchors[0];
             const a2 = anchors[1];
 
@@ -77,16 +77,15 @@ describe('focus manager', () => {
                 target: a2
             };
 
-            manager.handleTab(event);
+            manager.keyHandler(event);
 
             expect(document.activeElement).toEqual(a1);
         });
 
         it('should handle tabbing forward 1 element', () => {
-            const node = createNode();
-            const manager = new FocusManager(node);
+            const manager = new FocusManager(trapNode, controlNode);
 
-            const anchors = node.querySelectorAll('a');
+            const anchors = trapNode.querySelectorAll('a');
             const a2 = anchors[1];
             const a3 = anchors[2];
 
@@ -96,16 +95,15 @@ describe('focus manager', () => {
                 target: a2
             };
 
-            manager.handleTab(event);
+            manager.keyHandler(event);
 
             expect(document.activeElement).toEqual(a3);
         });
 
         it('should redirect to the last element when on first element', () => {
-            const node = createNode();
-            const manager = new FocusManager(node);
+            const manager = new FocusManager(trapNode, controlNode);
 
-            const anchors = node.querySelectorAll('a');
+            const anchors = trapNode.querySelectorAll('a');
             const a1 = anchors[0];
             const a4 = anchors[3];
 
@@ -115,16 +113,15 @@ describe('focus manager', () => {
                 target: a1
             };
 
-            manager.handleTab(event);
+            manager.keyHandler(event);
 
             expect(document.activeElement).toEqual(a4);
         });
 
         it('should redirect to the first element when on last element', () => {
-            const node = createNode();
-            const manager = new FocusManager(node);
+            const manager = new FocusManager(trapNode, controlNode);
 
-            const anchors = node.querySelectorAll('a');
+            const anchors = trapNode.querySelectorAll('a');
             const a1 = anchors[0];
             const a4 = anchors[3];
 
@@ -134,7 +131,7 @@ describe('focus manager', () => {
                 target: a4
             };
 
-            manager.handleTab(event);
+            manager.keyHandler(event);
 
             expect(document.activeElement).toEqual(a1);
         });